+++ src/index.html
@@ -12,14 +12,21 @@
 
 <body>
   <header>
-    <img class="menu" src="./icons/bars-solid.svg" />
+    <img class="menu" id="menu" src="./icons/bars-solid.svg" />
     <img src="./imagen/bannerTitulo.png" alt="">
   </header>
-  <nav>
-<div class="filtros"></div>
+  <nav id="menus" class="menus" >
+<div  class="filtros">
+<ul>
+  <li><a href="#">Localidad</a> </li>
+<li><a href="#">Genero</a> </li>
+<li><a href="#">Especie</a> </li>
+</ul>
+</div>
   </nav>
   <main>
       <section id="grid"></section>
+      
     </div>
   </main>
   <script src="main.js" type="module"></script>
